sip/zmq_logging_aggregator/requirements.txt
PyYAML>=3.12
pyzmq>=16.0.2
simplejson>=3.11.1
Flask>=0.12.2
